Rename package aws-ses to awsses
