+++ alerting/alert/type.go
@@ -5,7 +5,7 @@ package alert
 type Type string
 
 const (
-	// TypeAWSSES is the Type for the aws-ses alerting provider
+	// TypeAWSSES is the Type for the awsses alerting provider
 	TypeAWSSES Type = "aws-ses"
 
 	// TypeCustom is the Type for the custom alerting provider
