+++ alerting/provider/awsses/awsses.go
@@ -1,4 +1,4 @@
-package aws_ses
+package awsses
 
 import (
 	"fmt"
